Use CSS animations instead of transitions to show/hide the menu
When adding the `visibility: hidden` property we could not see the
transition while hiding the menu any more. So we had to find a way
to first run the effect/animation and after that apply the
`visibility: hidden` property.

Delaying `visibility` to run after the effect would not work with
`transitions` so for the animation we switched to CSS animations.